Rename matasano-crypto-solutions folder to matasano_crypto_solutions
