+++ matasano_crypto_solutions/6.txt
