+++ matasano_crypto_solutions/7.txt
