+++ matasano_crypto_solutions/8.txt
