+++ matasano_crypto_solutions/set2.py
@@ -1,6 +1,6 @@
 import os, random
 from collections import Counter
-from utils import (
+from matasano_crypto_solutions.utils import (
     base64_to_bytes,
     aes_ecb_decode,
     aes_ecb_encode,
@@ -54,38 +54,42 @@ def encryption_oracle(plaintext: bytes) -> bytes:
     to_encrypt = pad_with_pkcs7(before + plaintext + after, 16)
     enc_func = random.choice((aes_ecb_encode, aes_cbc_encode))
     if (enc_func == aes_cbc_encode):
-        print('oracle encrypted plaintext with cbc')
-        return enc_func(to_encrypt, random_aes_key(), random_aes_key())
+        return enc_func(to_encrypt, random_aes_key(), random_aes_key()), 'cbc'
     else:
-        print('oracle encrypted plaintext with aes')
-        return enc_func(to_encrypt, random_aes_key())
+        return enc_func(to_encrypt, random_aes_key()), 'ecb'
 
 
-def check_if_aes_or_cbc(text: bytes) -> str:
-    counts = Counter(split_into_groups(ciphertext, 16))
+def check_if_ecb_or_cbc(text: bytes) -> str:
+    counts = Counter(split_into_groups(text, 16))
     if counts.most_common(1)[0][1] > 1:
-        return 'aes'
+        return 'ecb'
     return 'cbc'
 
 
-print('Set 2')
-print('Challenge 9')
-res9 = pad_with_pkcs7(b'YELLOW SUBMARINE', 20)
-assert res9 == b'YELLOW SUBMARINE\x04\x04\x04\x04'
-print(res9)
+def main():
+    print('Set 2')
+    print('Challenge 9')
+    res9 = pad_with_pkcs7(b'YELLOW SUBMARINE', 20)
+    assert res9 == b'YELLOW SUBMARINE\x04\x04\x04\x04'
+    print(res9)
 
-print('Challenge 10')
-ciphertext10 = base64_to_bytes(get_file('10.txt'))
-password10 = b'YELLOW SUBMARINE'
-iv = b'\x00' * 16
-res10 = aes_cbc_decode(ciphertext10, password10, iv).decode('ascii')
-assert res10.startswith("I'm back and I'm ringin' the bell")
-print(res10)
-# Check that encrypting is the opposite of decrypting
-test_ciphertext10 = aes_cbc_encode(res10.encode('ascii'), password10, iv)
-assert test_ciphertext10 == ciphertext10
+    print('Challenge 10')
+    ciphertext10 = base64_to_bytes(get_file('10.txt'))
+    password10 = b'YELLOW SUBMARINE'
+    iv = b'\x00' * 16
+    res10 = aes_cbc_decode(ciphertext10, password10, iv).decode('ascii')
+    assert res10.startswith("I'm back and I'm ringin' the bell")
+    print(res10)
+    # Check that encrypting is the opposite of decrypting
+    test_ciphertext10 = aes_cbc_encode(res10.encode('ascii'), password10, iv)
+    assert test_ciphertext10 == ciphertext10
 
-print('Challenge 11')
-plaintext = text_to_bytes(res10)
-ciphertext = encryption_oracle(plaintext)
-print(check_if_aes_or_cbc(ciphertext))
+    print('Challenge 11')
+    plaintext = text_to_bytes(res10)
+    ciphertext, actual_answer = encryption_oracle(plaintext)
+    print(check_if_ecb_or_cbc(ciphertext))
+    print('Actual answer:', actual_answer)
+
+
+if __name__ == '__main__':
+    main()
