+++ matasano_crypto_solutions/utils.py
