+++ setup.py
@@ -9,7 +9,7 @@ setuptools.setup(name='matasano-crypto-solutions',
                  author='Ismail S',
                  author_email='ismail-s.no-reply@github.com',
                  url='https://github.com/ismail-s/Matasano-Crypto-Solutions',
-                 py_modules=['matasano-crypto-solutions'],
+                 py_modules=['matasano_crypto_solutions'],
                  install_requires=open('requirements.txt').read().split(),
                  license='MIT License',
                  zip_safe=True)
